kuring-71 [수정] 알림 화면 코드를 모듈로 옮김
