+++ feature/my_notification/src/main/java/com/ku_stacks/ku_ring/my_notification/NotificationActivity.kt
@@ -1,4 +1,4 @@
-package com.ku_stacks.ku_ring.ui.my_notification
+package com.ku_stacks.ku_ring.my_notification
 
 import android.app.Activity
 import android.content.Intent
@@ -8,11 +8,10 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.ku_stacks.ku_ring.R
-import com.ku_stacks.ku_ring.databinding.ActivityNotificationBinding
+import com.ku_stacks.ku_ring.my_notification.databinding.ActivityNotificationBinding
+import com.ku_stacks.ku_ring.my_notification.mapper.toWebViewNotice
+import com.ku_stacks.ku_ring.my_notification.ui_model.PushContentUiModel
 import com.ku_stacks.ku_ring.thirdparty.firebase.analytics.EventAnalytics
-import com.ku_stacks.ku_ring.ui.my_notification.ui_model.PushContentUiModel
-import com.ku_stacks.ku_ring.ui.notice_webview.NoticeWebActivity
 import com.ku_stacks.ku_ring.ui_util.KuringNavigator
 import com.ku_stacks.ku_ring.ui_util.makeDialog
 import com.yeonkyu.HoldableSwipeHelper.HoldableSwipeHandler
@@ -121,7 +120,7 @@ class NotificationActivity : AppCompatActivity() {
     }
 
     private fun startNoticeActivity(pushContent: PushContentUiModel) {
-        NoticeWebActivity.start(this, pushContent)
+        navigator.navigateToNoticeWeb(this, pushContent.toWebViewNotice())
     }
 
     private fun startMainActivity() {
