+++ feature/my_notification/src/main/java/com/ku_stacks/ku_ring/my_notification/NotificationAdapter.kt
@@ -1,19 +1,18 @@
-package com.ku_stacks.ku_ring.ui.my_notification
+package com.ku_stacks.ku_ring.my_notification
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.ku_stacks.ku_ring.R
-import com.ku_stacks.ku_ring.databinding.ItemDateBinding
-import com.ku_stacks.ku_ring.databinding.ItemNotificationBinding
-import com.ku_stacks.ku_ring.ui.my_notification.diff_callback.NotificationDiffCallback
-import com.ku_stacks.ku_ring.ui.my_notification.ui_model.PushContentUiModel
-import com.ku_stacks.ku_ring.ui.my_notification.ui_model.PushDataUiModel
-import com.ku_stacks.ku_ring.ui.my_notification.ui_model.PushDateHeaderUiModel
-import com.ku_stacks.ku_ring.ui.my_notification.viewholder.DateViewHolder
-import com.ku_stacks.ku_ring.ui.my_notification.viewholder.NotificationViewHolder
+import com.ku_stacks.ku_ring.my_notification.databinding.ItemDateBinding
+import com.ku_stacks.ku_ring.my_notification.databinding.ItemNotificationBinding
+import com.ku_stacks.ku_ring.my_notification.diff_callback.NotificationDiffCallback
+import com.ku_stacks.ku_ring.my_notification.ui_model.PushContentUiModel
+import com.ku_stacks.ku_ring.my_notification.ui_model.PushDataUiModel
+import com.ku_stacks.ku_ring.my_notification.ui_model.PushDateHeaderUiModel
+import com.ku_stacks.ku_ring.my_notification.viewholder.DateViewHolder
+import com.ku_stacks.ku_ring.my_notification.viewholder.NotificationViewHolder
 
 class NotificationAdapter(
     private val itemClick: (PushContentUiModel) -> Unit,
@@ -31,12 +30,14 @@ class NotificationAdapter(
                 val binding = ItemNotificationBinding.bind(view)
                 NotificationViewHolder(binding, itemClick)
             }
+
             NOTIFICATION_DATE -> {
                 val view =
                     LayoutInflater.from(parent.context).inflate(R.layout.item_date, parent, false)
                 val binding = ItemDateBinding.bind(view)
                 DateViewHolder(binding)
             }
+
             else -> throw Exception("no such viewType : $viewType")
         }
     }
@@ -49,6 +50,7 @@ class NotificationAdapter(
                 holder.bind(item as PushContentUiModel)
                 onBindItem(item)
             }
+
             is DateViewHolder -> {
                 holder.bind(item as PushDateHeaderUiModel)
             }
