+++ feature/my_notification/src/main/java/com/ku_stacks/ku_ring/my_notification/diff_callback/NotificationDiffCallback.kt
@@ -1,9 +1,9 @@
-package com.ku_stacks.ku_ring.ui.my_notification.diff_callback
+package com.ku_stacks.ku_ring.my_notification.diff_callback
 
 import androidx.recyclerview.widget.DiffUtil
-import com.ku_stacks.ku_ring.ui.my_notification.ui_model.PushContentUiModel
-import com.ku_stacks.ku_ring.ui.my_notification.ui_model.PushDataUiModel
-import com.ku_stacks.ku_ring.ui.my_notification.ui_model.PushDateHeaderUiModel
+import com.ku_stacks.ku_ring.my_notification.ui_model.PushContentUiModel
+import com.ku_stacks.ku_ring.my_notification.ui_model.PushDataUiModel
+import com.ku_stacks.ku_ring.my_notification.ui_model.PushDateHeaderUiModel
 
 class NotificationDiffCallback : DiffUtil.ItemCallback<PushDataUiModel>() {
     override fun areItemsTheSame(oldItem: PushDataUiModel, newItem: PushDataUiModel): Boolean {
