+++ feature/my_notification/src/main/java/com/ku_stacks/ku_ring/my_notification/ui_model/PushUiModel.kt
@@ -1,4 +1,4 @@
-package com.ku_stacks.ku_ring.ui.my_notification.ui_model
+package com.ku_stacks.ku_ring.my_notification.ui_model
 
 sealed class PushDataUiModel
