+++ feature/my_notification/src/main/res/layout/activity_notification.xml
